+++ service/internal/api/conesearch_handlers.go
@@ -1,3 +1,17 @@
+// Copyright 2024-2025 Diego Rodriguez Mancini
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//	http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package api
 
 import (
@@ -136,7 +150,7 @@ func (api *API) conesearch(c *gin.Context) {
 }
 
 func isEmptyResult(result any) bool {
-	// WARN: in this layer there should not be any references to repository
+	// WARN: in this layer there should not be any references to repository, that should be done in service
 	switch result.(type) {
 	case []repository.Mastercat:
 		return len(result.([]repository.Mastercat)) == 0
