+++ service/internal/api/ping_test.go
@@ -12,22 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package httpservice_test
+package api_test
 
-// import (
-	// "net/http"
-	// "net/http/httptest"
-	// "testing"
+import (
+	"net/http"
+	"net/http/httptest"
+	"testing"
 
-	// "github.com/stretchr/testify/require"
-// )
+	"github.com/stretchr/testify/require"
+)
 
-// func TestPing(t *testing.T) {
+func TestPing(t *testing.T) {
 
-	// w := httptest.NewRecorder()
-	// req, _ := http.NewRequest("GET", "/ping", nil)
-	// router.ServeHTTP(w, req)
+	w := httptest.NewRecorder()
+	req, _ := http.NewRequest("GET", "/ping", nil)
+	router.ServeHTTP(w, req)
 
-	// require.Equal(t, http.StatusOK, w.Code)
-	// require.Equal(t, "pong", w.Body.String())
-// }
+	require.Equal(t, http.StatusOK, w.Code)
+	require.Equal(t, "pong", w.Body.String())
+}
