+++ service/internal/api/routes.go
@@ -1,7 +1,22 @@
+// Copyright 2024-2025 Diego Rodriguez Mancini
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//	http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package api
 
 import (
 	"net/http"
+	"os"
 
 	"github.com/gin-gonic/gin"
 	swaggerFiles "github.com/swaggo/files"
@@ -9,8 +24,11 @@ import (
 )
 
 func (api *API) SetupRoutes(r *gin.Engine) {
+	if r == nil {
+		panic("api: gin engine cannot be nil")
+	}
 	r.Use(gin.Recovery())
-	if api.getEnv("USE_LOGGER") != "" {
+	if os.Getenv("USE_LOGGER") != "" {
 		r.Use(gin.Logger())
 	}
